refactor: revert to pre-production transfers/
